+++ src/ScrumBoardItBundle/Resources/views/Default/index.html.twig
@@ -4,7 +4,7 @@
 	<div class="header">
 		<img src="{{ asset('bundles/scrumboardit/images/header.png') }}" />
 		<div class="log">
-			<img class="logImg" src="{{ app.user.imgUrl }}" /> <span>{{
+			<img class="logImg" src="{{ app.user.imgUrl }}" height="24" width="24"/> <span>{{
 				app.user.displayName }}</span><br /> <a href="./logout"><span
 				class="glyphicon glyphicon-off">Déconnexion</span></a>
 		</div>
