+++ src/ScrumBoardItBundle/Resources/views/Security/login.html.twig
@@ -24,7 +24,7 @@
 						<label>Se connecter à :</label>
 						<select name="_api" class="form-control">
 							<option value="jira">Jira</option>
-							<option value="github">GitHub</option>
+							<option value="github" selected>GitHub</option>
 						</select>
 					</div>
 					<div class="form-group">
