+++ src/ScrumBoardItBundle/Services/ApiGitHub.php
@@ -1,0 +1,48 @@
+<?php
+namespace ScrumBoardItBundle\Services;
+
+use ScrumBoardItBundle\Services\AbstractApi;
+use Symfony\Component\HttpFoundation\Request;
+
+class ApiGitHub extends AbstractApi
+{
+    /**
+     *
+     * {@inheritdoc}
+     *
+     */
+    public function getProjects()
+    {
+        
+    }
+    
+    /**
+     *
+     * {@inheritdoc}
+     *
+     */
+    public function searchIssues($searchFilters = null)
+    {
+        
+    }
+    
+    /**
+     *
+     * {@inheritdoc}
+     *
+     */
+    public function getSelectedIssues(Request $request, $selected)
+    {
+        
+    }
+    
+    /**
+     *
+     * {@inheritdoc}
+     *
+     */
+    public function getSearchFilters(Request $request)
+    {
+        
+    }
+}
